GPII-3848: refactoring qss tests
